Remove some unnecessary async annotations (#22)
`performCompleteDiarization` and `cleanup` do not perform any
asynchronous work and it is difficult to imagine any reason they would
ever want to.

Fundamentally, performing diarization is not an asynchronous operation -
it's just a calculation, and you don't need to wait arbitrarily long on
external systems such as the network or filesystem.

There are some async operations you can build around diarization, such
as processing an async stream of samples, or concurrently processing
multiple audio buffers (as shown in `compareSpeakers` which now uses
`async let`), and loading the model makes sense to do asynchronously,
but the computation itself is synchronous and it's important to make it
available in synchronous contexts.

Co-authored-by: Brandon Weng <18161326+BrandonWeng@users.noreply.github.com>