[Performance] Eliminate some copies of the sample buffer (#21)
The sample buffer can be rather large, and it would be nice to avoid
copying it.

---

In `performCompleteDiarization`, there is a loop which splits the
provided sample array in to 10-second chunks. Unfortunately, because
later functions require an `Array` as input, each of these chunks are
copies from the provided array in to fresh storage.

We can change the other functions to support `ArraySlice`s as input
instead (any Array can trivially be converted to an ArraySlice using
`array[...]`). Some small adjustments are required because they
sometimes assume the input starts at index 0, which when using slices it
now might not.

Ultimately, the sample buffer might be copied in
`processChunkWithSpeakerTracking` if it needs padding. I've rewritten
that copy to minimise reallocations, but in a later PR I'd like to have
it provided externally, so it can be hoisted out of the loop in
`performCompleteDiarization` and can be allocated once and reused.

---

Additionally, in `getEmbedding`, the audio chunk is copied in to a
tensor of shape `(numspeakers, 10 * 16_000)`, where each row of
`numspeakers` is an identical copy of the audio chunk. That data then
gets copied in to the `waveformArray` and is never used again.

This doesn't actually allocate a huge amount of storage since the sample
buffer is still stored indirectly, but it's still an array that isn't
needed. We can just access the value from the single-row audio chunk
directly.

Co-authored-by: Alex <36247722+Alex-Wengg@users.noreply.github.com>