Add a models parameter to DiarizerManager.initialize (#28)
Only the last commit is new.

Builds on #25 by adding a `models` parameter to
`DiarizerManager.initialize`.

The `DiarizerModels` type gives users APIs to download models, including
to custom directories, and load models that are bundled with the app or
downloaded from elsewhere. Models can also be loaded with user-specified
options, such as the compute units to use.

This unlocks some new use-cases but comes with some minor API changes.
IMO, it is actually quite nice to let users know that initialising the
diarization manager might involve a download, so I feel the extra
clarity is worth the extra characters.

The `DiarizerModels` APIs make the previous
`DiarizerManager.downloadModels` and
`DiarizerConfig.modelCacheDirectory` APIs obsolete, so they have also
been removed.