Refactor diarization model loading in to DiarizationModels type (#25)
Firstly, thank you very much for creating this library. We are currently
exploring whether it would be suitable for an iOS app, but there are a
few very minor alterations we need. We think these alterations should
not be controversial and hopefully can deliver some benefits to all
users of the library:

1. Support for pre-downloaded models. For the use-case we have in mind,
downloading models from huggingface would not be okay. Luckily the
models are only 36MB, so shipping them with the app or managing our
downloading from our private server using some other mechanism is
reasonable.

2. Support for maintaining the speaker embedding database across
diarization calls. Basically just adding a version of the API which
takes the speaker DB as an `inout` parameter.

This PR is part of the first of these.


----


The `DiarizationManager` currently performs two major functions:

1. model management, and
2. diarization and related functions (e.g. `compareSpeakers`)

This PR splits moves the model loading in to a new type,
`DiarizationModels`, and leaves the calculations in
`DiarizationManager`.

It is mostly compatible with the existing public API, with one minor
change: `ModelPaths` members are now URLs instead of Strings. There was
a lot of URL->String->URL conversion in the previous code, but since the
system APIs all take URLs, it is probably better if we just stick with
URLs everywhere.